Refactored platform service to use blueprints and breakdown large main file
