+++ platform-service/src/unstract/platform_service/env.py
@@ -1,0 +1,27 @@
+import os
+
+from unstract.platform_service.utils import EnvManager
+
+
+class Env:
+    INVALID_ORGANIZATOIN = "Invalid organization"
+    INVALID_PAYLOAD = "Bad Request / No payload"
+    BAD_REQUEST = "Bad Request"
+    REDIS_HOST = EnvManager.get_required_setting("REDIS_HOST")
+    REDIS_PORT = int(EnvManager.get_required_setting("REDIS_PORT", 6379))
+    REDIS_USERNAME = os.environ.get("REDIS_USERNAME")
+    REDIS_PASSWORD = os.environ.get("REDIS_PASSWORD")
+    PG_BE_HOST = os.environ.get("PG_BE_HOST")
+    PG_BE_PORT = int(os.environ.get("PG_BE_PORT", 5432))
+    PG_BE_USERNAME = os.environ.get("PG_BE_USERNAME")
+    PG_BE_PASSWORD = os.environ.get("PG_BE_PASSWORD")
+    PG_BE_DATABASE = os.environ.get("PG_BE_DATABASE")
+    ENCRYPTION_KEY = EnvManager.get_required_setting("ENCRYPTION_KEY")
+    MODEL_PRICES_URL = EnvManager.get_required_setting("MODEL_PRICES_URL")
+    MODEL_PRICES_TTL_IN_DAYS = int(
+        EnvManager.get_required_setting("MODEL_PRICES_TTL_IN_DAYS")
+    )
+    MODEL_PRICES_FILE_PATH = EnvManager.get_required_setting("MODEL_PRICES_FILE_PATH")
+
+
+EnvManager.raise_for_missing_envs()
